Move LinkedList, RingBuffer, & Bitset to XS
