+++ SConstruct
@@ -102,6 +102,7 @@ env = Environment(
         'VER_NAME': '\\"$VERSION_NAME\\"',
     },
     CPPPATH=[
+        '#',
         '#Kernel',
         '#Thirdparty',
         '#Libraries',
@@ -141,15 +142,17 @@ if 'docs' not in COMMAND_LINE_TARGETS:
     kernel_targets_debug = []
     kernel_targets_release = []
     for target_env in kernel_environments:
-        libxs = target_env.SConscript(
-            'XS/SConscript',
-            variant_dir='$BUILD_DIR/libxs',
-            duplicate=0,
-            exports={
-                'env': target_env
-            },
-        )
-        Default(libxs)
+        # TODO: Uncomment as soon as there's source files in XS
+        #
+        # libxs = target_env.SConscript(
+        #     'XS/SConscript',
+        #     variant_dir='$BUILD_DIR/libxs',
+        #     duplicate=0,
+        #     exports={
+        #         'env': target_env
+        #     },
+        # )
+        # Default(libxs)
         liballoc = target_env.SConscript(
             'Libraries/liballoc/SConscript',
             variant_dir='$BUILD_DIR/liballoc',
@@ -180,12 +183,12 @@ if 'docs' not in COMMAND_LINE_TARGETS:
         )
         Default(image)
 
-        kernel_targets_all.extend([libxs, liballoc, kernel, image])
+        kernel_targets_all.extend([liballoc, kernel, image])
 
         # Add targets to kernel_targets_[MODE] list
         target_list_name = 'kernel_targets_' + target_env['MODE'].lower()
         targets_list = globals()[target_list_name]
-        targets_list.extend([libxs, liballoc, kernel, image])
+        targets_list.extend([liballoc, kernel, image])
 
     # Mode specific kernel targets
     env.Alias('kernel-debug', kernel_targets_debug)
@@ -235,6 +238,7 @@ env = Environment(
         '-ftest-coverage',
     ],
     CPPPATH=[
+        '#',
         '#Tests',
         '#Kernel',
         '#Thirdparty',
